Moved HTTP cerver auth related methods to dedicated sources
